my broke code
