--- app/views/templates/hello.hbs
@@ -1,8 +1,0 @@
-<script id="CDTemplate" type="text/x-handlebars-template">
-      <h1>{{ name }}</h1>
-      <h4>{{ artist }}</h4>
-      <p>{{ description }}</p>
-      <ul>{{ tracks }}
-
-      </ul>
-    </script>
